requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements/requirements.txt requirements/requirements.in
#
amqp==5.1.1
    # via kombu
asgiref==3.7.2
    # via django
async-timeout==4.0.2
    # via redis
billiard==3.6.4.0
    # via celery
celery==5.2.7
    # via
    #   -r requirements/requirements.in
    #   django-celery-beat
    #   django-celery-results
    #   flower
    #   pytest-celery
cerberus==1.3.4
    # via -r requirements/requirements.in
certifi==2023.5.7
    # via requests
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
coreapi==2.3.3
    # via drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
coverage[toml]==7.2.5
    # via
    #   -r requirements/requirements.in
    #   pytest-cov
cron-descriptor==1.4.0
    # via django-celery-beat
django==4.2.2
    # via
    #   -r requirements/requirements.in
    #   django-celery-beat
    #   django-celery-results
    #   django-filter
    #   django-jazzmin
    #   django-redis
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-yasg
    #   model-bakery
django-celery-beat==2.5.0
    # via -r requirements/requirements.in
django-celery-results==2.5.1
    # via -r requirements/requirements.in
django-dotenv==1.4.2
    # via -r requirements/requirements.in
django-filter==23.2
    # via -r requirements/requirements.in
django-jazzmin==2.6.0
    # via -r requirements/requirements.in
django-redis==5.3.0
    # via -r requirements/requirements.in
django-timezone-field==5.0
    # via django-celery-beat
djangorestframework==3.14.0
    # via
    #   -r requirements/requirements.in
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework-simplejwt==5.2.2
    # via -r requirements/requirements.in
drf-yasg==1.21.5
    # via -r requirements/requirements.in
exceptiongroup==1.1.1
    # via pytest
flake8==6.0.0
    # via -r requirements/requirements.in
flower==1.2.0
    # via -r requirements/requirements.in
humanize==4.6.0
    # via flower
idna==3.4
    # via requests
inflection==0.5.1
    # via drf-yasg
iniconfig==2.0.0
    # via pytest
itypes==1.2.0
    # via coreapi
jinja2==3.1.2
    # via coreschema
kombu==5.3.1
    # via celery
markupsafe==2.1.3
    # via jinja2
mccabe==0.7.0
    # via flake8
model-bakery==1.11.0
    # via -r requirements/requirements.in
packaging==23.1
    # via
    #   drf-yasg
    #   pytest
pluggy==1.0.0
    # via pytest
prometheus-client==0.17.0
    # via flower
prompt-toolkit==3.0.38
    # via click-repl
psycopg2-binary==2.9.6
    # via -r requirements/requirements.in
pycodestyle==2.10.0
    # via flake8
pyflakes==3.0.1
    # via flake8
pyjwt==2.7.0
    # via djangorestframework-simplejwt
pytest==7.3.2
    # via
    #   -r requirements/requirements.in
    #   pytest-cov
    #   pytest-django
    #   pytest-mock
pytest-celery==0.0.0
    # via -r requirements/requirements.in
pytest-cov==4.1.0
    # via -r requirements/requirements.in
pytest-django==4.5.2
    # via -r requirements/requirements.in
pytest-mock==3.10.0
    # via -r requirements/requirements.in
python-crontab==2.7.1
    # via django-celery-beat
python-dateutil==2.8.2
    # via python-crontab
pytz==2023.3
    # via
    #   celery
    #   django-timezone-field
    #   djangorestframework
    #   drf-yasg
    #   flower
redis==4.5.4
    # via
    #   -r requirements/requirements.in
    #   django-redis
requests==2.31.0
    # via
    #   -r requirements/requirements.in
    #   coreapi
ruamel-yaml==0.17.32
    # via drf-yasg
ruamel-yaml-clib==0.2.7
    # via ruamel-yaml
six==1.16.0
    # via python-dateutil
sqlparse==0.4.4
    # via django
tomli==2.0.1
    # via
    #   coverage
    #   pytest
tornado==6.3.2
    # via flower
typing-extensions==4.6.3
    # via asgiref
tzdata==2023.3
    # via django-celery-beat
uritemplate==4.1.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==2.0.3
    # via requests
uwsgi==2.0.21
    # via -r requirements/requirements.in
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.6
    # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
